feat: add webdav locks backend

+++ database/migrations/2026_06_03_121217_create_dav_locks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->unsignedInteger('created');
             $table->string('token')->unique();
             $table->unsignedTinyInteger('scope');
-            $table->unsignedTinyInteger('depth');
+            $table->integer('depth');
             $table->text('uri');
             $table->timestamps();
